Add files via upload
This is all the data I uploaded to create the JoBot. It includes the writing from Jo's four diaries (March 1880- December 1881, July 1883- August 1885, September 1885- January 1892, and February 1892- May 1897), the letters she wrote to Theo during their engagement (December 1888- April 1889), and the introduction she wrote for the collection of Vincent and Theo's letters.